refactor: LLM 모듈 common/llm/으로 승격 및 Bot-API 캐싱 추가
- LLM 추상화 레이어(LlmProvider, GeminiLlmProvider, LlmModule)를
  voice-analytics/infrastructure/llm/ → common/llm/으로 이동하여
  다른 도메인에서도 재사용 가능하도록 승격
- Bot-API의 analyze, community-health 엔드포인트에 Redis 캐싱 추가 (TTL 30분)
- 문서 정합성 수정: master plan C.7 완료 반영, PRD 경로 현행화,
  모델 기본값 gemini-2.5-flash 반영
