+++ apps/api/src/voice-analytics/voice-analytics.module.ts
@@ -6,12 +6,12 @@ import { AuthModule } from '../auth/auth.module';
 import { VoiceCoPresencePairDailyOrm } from '../channel/voice/co-presence/infrastructure/voice-co-presence-pair-daily.orm-entity';
 import { VoiceDailyOrm } from '../channel/voice/infrastructure/voice-daily.orm-entity';
 import { VoiceRedisRepository } from '../channel/voice/infrastructure/voice-redis.repository';
+import { LlmModule } from '../common/llm/llm.module';
 import { GatewayModule } from '../gateway/gateway.module';
 import { MocoHuntingDailyOrmEntity as MocoHuntingDaily } from '../newbie/infrastructure/moco-hunting-daily.orm-entity';
 import { VoiceAiAnalysisService } from './application/voice-ai-analysis.service';
 import { VoiceAnalyticsService } from './application/voice-analytics.service';
 import { VoiceNameEnricherService } from './application/voice-name-enricher.service';
-import { LlmModule } from './infrastructure/llm/llm.module';
 import { VoiceAnalyticsController } from './presentation/voice-analytics.controller';
 import { BadgeScheduler } from './self-diagnosis/application/badge.scheduler';
 import { BadgeService } from './self-diagnosis/application/badge.service';
